Use concerns for Synchronizations and IdMap as well
