+++ app/models/maestrano/connector/rails/concerns/id_map.rb
@@ -1,0 +1,8 @@
+module Maestrano::Connector::Rails::Concerns::IdMap
+  extend ActiveSupport::Concern
+
+  included do
+    belongs_to :organization
+    serialize :metadata, Hash
+  end
+end
