+++ app/models/maestrano/connector/rails/id_map.rb
@@ -1,6 +1,5 @@
 module Maestrano::Connector::Rails
   class IdMap < ActiveRecord::Base
-    belongs_to :organization
-    serialize :metadata, Hash
+    include Maestrano::Connector::Rails::Concerns::IdMap
   end
 end
